+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # 2.x branch
 ## 2.18 branch
 ### 2.19.0
-* `config()` global function becomes `getConfig()`.
+* `config()` global function becomes `getConfig()` The function accepts a second
+    parameter as the default value in case the configuration is empty.
 
 ### 2.18.1
 * `BannersController::upload()` and `PhotosController::upload()` methods return